+++ packages/react/src/web-widget.tsx
@@ -84,6 +84,7 @@ export /*#__PURE__*/ function defineWebWidget(
   options: DefineWebWidgetOptions
 ) {
   options.renderTarget = "light"; // TODO shadow
+  options.loading = "lazy";
   return function WebWidgetSuspense({
     children,
     fallback,