+++ packages/vue/src/web-widget.ts
@@ -93,6 +93,7 @@ export /*#__PURE__*/ function defineWebWidget(
   options: DefineWebWidgetOptions
 ) {
   options.renderTarget = "light"; // TODO shadow
+  options.loading = "lazy";
   return /*#__PURE__*/ defineComponent({
     name: "WebWidgetSuspense",
     props: {